// src/modules/workspaces/workspace.enums.ts
export enum WorkspaceType {
  Personal = 'personal',
  Team = 'team',
}

export enum WorkspaceStatus {
  Active = 'active',
  Suspended = 'suspended',
  Archived = 'archived',
}

export enum WorkspaceRole {
  Owner = 'owner',
  Admin = 'admin',
  Billing = 'billing',
  Member = 'member',
  Viewer = 'viewer',
}

export enum WorkspaceMemberStatus {
  Active = 'active',
  Suspended = 'suspended',
}